pyproject.toml
[project]
name = "odata-avps-drhfpnc"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "marker-pdf>=1.10.2",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "requests>=2.33.1",
]
